+++ test/exop_data/common_filters_test.exs
@@ -27,5 +27,17 @@ defmodule ExopData.CommonFiltersTest do
       assert Enum.count(check_list_nils) == 0
       assert Enum.count(check_list) == Enum.count(check_list_ints)
     end
+
+    property "with 'false' by default" do
+      contract = [%{name: :a, opts: [required: true, type: :integer]}]
+
+      check_list = contract |> ExopData.generate() |> Enum.take(1000)
+
+      check_list_ints = Enum.filter(check_list, fn %{a: a} -> is_integer(a) end)
+      check_list_nils = Enum.filter(check_list, fn %{a: a} -> is_nil(a) end)
+
+      assert Enum.count(check_list_nils) == 0
+      assert Enum.count(check_list) == Enum.count(check_list_ints)
+    end
   end
 end